Encapsulate actual-mic-in-use bookkeeping in AudioManager, freeing the transport from doing yet another bit of bookkeeping.
As a bonus, this solves a bug where we nondeterministically end up stuck using a Bluetooth device right after connecting it, even if the user has an explicit device preference for a different device. The reason has to do with the Bluetooth device triggering an additional route change event some time *after* the route change event reporting the device’s availability. Because that additional route change doesn't add or remove devices, it was previously being ignored.